Always bucket video metadata for DB I/O. Establish //media/capabilities.
This fixes a bug where we were not properly bucketing on the query-side.
We now bucket for all reads and writes.

Bucketting is enforced by the creation of a "bucketted"
VideoDecodeStatsDB::VideoDescKey. The key constructor is now private, so
all keys must be created via the bucketting path.

Bucketing also occurs in the VideoDecodeStatsReporter. This is mostly to
ensure that we don't start new stats records for slight changes to metadata
that wouldn't otherwise cause stats to be written to a new bucket.

Bug: 782707
Change-Id: Ib8c1ded93ec8fe31014d74e45963863ebdb45788
Reviewed-on: https://chromium-review.googlesource.com/764572
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#516822}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9e669479fcec5e72c7534b204cb60893b12895f1