media: Add additional trace events for MojoCdmFileIO
Adds trace events for opening the file, as well as when Write() and
WriteDone() are called to understand the impact of opening a temporary
file and renaming it after it's written to. Also adds fileName to the
event data saved.

BUG=774160

Change-Id: Id0476d028865802e1ddda22eac6620cd6f16f106
Reviewed-on: https://chromium-review.googlesource.com/775654
Commit-Queue: John Rummell <jrummell@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#517313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f45580a8857e186a038745049e595bd31280f1cf