media: Remove WeakPtr usage on MojoCdmServiceContext
Currently we use StrongBindingSet for mojo interface implementations
that needs access to MojoCdmServiceContext. Hence the lifetime of those
objects are determined and we should be able to use raw pointers instead
of weak pointers.

In MojoCdmService, MojoAudioDecoderService and MojoVideoDecoderService
we assume the MojoCdmServiceContext always exists.

In MojoRendererService, where it is currently also used outside of
MediaService (see RenderFrameHostImpl), it's possible that
MojoCdmServiceContext can be null, and we have explicit check on that.

BUG=786736
TEST=Existing tests pass.

Change-Id: Idd7e0a1bb24d109c7a17607e3fd856506432e9ee
Reviewed-on: https://chromium-review.googlesource.com/780046
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518568}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d44c932dd455b5f12346dfdb2ab60a56a34c442d