Fix CodecSupportTest_av1 when proprietary_codecs=false
This test was not actually testing the right thing, by default
video/mp4 and video/webm with an unknown codec are reported as
"probably" being supported. Fix this so that av1 is recognized
and return "" if support is not compiled or enabled.

BUG=787514
TEST=updated browsertest

Change-Id: I14a62d5789b1def48413989eb859dcd1a878294b
Reviewed-on: https://chromium-review.googlesource.com/783330
Reviewed-by: Johann Koenig <johannkoenig@google.com>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#519402}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b536391fe46113bac96028988b6fdd6447b4d215