Manually release MediaCrypto plugin
After the end of a stream, we now manually release the MediaCrypto
plugin, so that it can be cleaned up on the TZ-side. If this is not
done, then we risk the chance of GC not cleaning up before the TZ heap
overflows.

TEST=Switch HBO Go cast stream multiple times, and watch the TZ heap
BUG=b:69682820

Change-Id: I6fda96b74525533b64638d859cb79a4aa6375248
Reviewed-on: https://chromium-review.googlesource.com/790570
Commit-Queue: Aidan Wolter <awolter@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#519911}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7ab2c15c67dd31e2f7058e2ae332640d8ebd50a4