Increase VideoFrame stride alignment to 32; required by ffmpeg.
Per code inspection and commentary from ffmpeg developers, we
need to be using a stride that is aligned to 32 for software
video frames.

It seems at somepoint we might be able to use av_cpu_max_align(),
but even ffmpeg doesn't follow this internally and has exceptions
for when H264 content is used, so just leave a TODO for now.

BUG=773673
TEST=no more ubsan failure.

Change-Id: I012bf7ca5531559df63061fc1edc9fdeaebf4c08
Reviewed-on: https://chromium-review.googlesource.com/801775
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#520727}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a648b776f80a5039f7a78558b3acfb68975071dc