[media] Create channel for cross-process VideoFrame release.
This CL defines a new interface, VideoFrameReleaser, and adds a channel
in MojoVideoDecoder to use it, in a way that can outlive the
VideoDecoder.

On the renderer side, MojoVideoFrameReleaser is ref counted, and will
live as long as there are outstanding VideoFrames.

On the gpu side, VideoFrameHandleReleaserImpl maintains a map of
outstanding VideoFrames, and will live until the map is empty.

The old path, with special methods on mojom::VideoDecoder, is removed.
VideoDecoder implementations can now attach regular mailbox release
callbacks to their VideoFrames and expect normal results.

Bug: 737220, 733828
Change-Id: I3c26e1e98c61f180246815fcc79f6731d8a26ed4
Reviewed-on: https://chromium-review.googlesource.com/762104
Commit-Queue: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#521831}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b795cd5ca7a49a8e5da74ff14abf4613ab50d59c