[Video/image capture] Revert refactoring CL that triggers intermittent freeze on Android
Refactoring CL being reverted:
https://chromium.googlesource.com/chromium/src/+/c69a479170be23ec6fed72bf863b18c44cdbd694

This also removes a block of code about "focus adjustments" which was added on
top of the refactoring and requires it.

The revert has been verified manually to fix the symptoms on a set of 20+ devices.

Bug: 793859
Change-Id: I9758b5dd6391682bee57a13756188d39cef20218
Reviewed-on: https://chromium-review.googlesource.com/820674
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#527815}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 68b440535588b20776bdf50524ecc9168d7c5e8f