Mock WebMediaPlayerClient.
This CL mocks the WebMediaPlayerClient (previously a fake) in
unittesting to confirm that the correct web_layer is being passed to
it. I've included a few other expect calls, but since some of the
call patterns are highly variable based on commandline args, I
demoted this to a NiceMock.

Bug: 755880
Change-Id: I1cf844701a115551ff49d104ec61d39da1774eaa
Reviewed-on: https://chromium-review.googlesource.com/804699
Commit-Queue: CJ DiMeglio <lethalantidote@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#528832}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ef15b62a4bbedf654f86fb5149ac23026c23af61