media/cast: Fix performance overlay frame copy corruption.
The performance metrics debug overlay broke recently. The root cause is
that when a copy of the VideoFrame is being made,

than the source frame. This change adds logic to copy the rows of the
frame one-by-one, to account for the different strides of the source and
destination frames.

VideoFrame: :CreateFrame() is allocating a frame with a stride larger
Change-Id: Ia32a74ca6f3279412bbbb81946070a2596ba6e73
Reviewed-on: https://chromium-review.googlesource.com/874515
Commit-Queue: Yuri Wiitala <miu@chromium.org>
Reviewed-by: Xiangjun Zhang <xjz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#530760}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3881e7ba4541a0a9a276144edb6a674758bd6a7b