Remove content deps of AudioSyncReader and AudioInputSyncWriter.
They will be moved to media in a future CL since they need to be used
from the audio service. A side-effect of this CL is that logs from the
SyncReader/Writer are prefixed with the stream_id, like other logs are.

Bug: 803102
Change-Id: Iae2876033e4172066d5a0a3b1322092f870dc32d
Reviewed-on: https://chromium-review.googlesource.com/878627
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Max Morin <maxmorin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#531849}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 13ddfd3252448db55a45a5f9fce07f1f664d282b