media: Report origin when library CDM crashes on decrypt and/or decode
When the CDM crashes in Decyrpt(), DecryptAndDecodeAudio() or
DecryptAndDecodeVideo(), we'll report the security origin in the crash
report for debugging purposes.

Tested with --enable-crash-reporter-for-testing, which will upload
a crash report and print a crash ID to stdout. Then confirmed that
we have the origin in the crash report.

See crash a689d5ab482dbb4a for an example.

BUG=785139

Change-Id: I72844532d5f3d5b03ce7401eab89e53d9509b169
Reviewed-on: https://chromium-review.googlesource.com/915058
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#536448}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8b6f44980094ddd78b25624c9bbc73652383b95