Shuts off SurfaceLayerForVideo in disable gpu case.
This is a temp fix disabling Surface Layers for video in the case
of disabling the gpu. It does not fix the issue, but make the feature
more stable for testing until we can in fact debug the problem.

Bug: 807840
Change-Id: Iada63e5b2b4c27988a75b56d51851a26f793a7be
Reviewed-on: https://chromium-review.googlesource.com/927223
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Commit-Queue: CJ DiMeglio <lethalantidote@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#538501}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db88a6705da3280e663a675bb415ed22d8a7de3f