bindings: Prepare V8ContextSnapshot on Android
This CL makes it possible to create and use V8 context snapshot
files on Android, if a developer puts "use_v8_context_snapshot=true"

The V8 context snapshot feature is currently supported and enabled on
Windows, MacOSX, and Linux (not ChromeOS), and its Finch experiment
showed 20% reduction of time to create V8 contexts.
On local Android machine, this feature shows 50% time recution.

Beside it, in order to enable the feature on cross-build platforms
like Android, we have to build Blink twice.
So this CL does not flip the flag |use_v8_context_snapshot|.

FYI, some numbers with the feature are written in the next CL;
https://chromium-review.googlesource.com/c/chromium/src/+/923623


TBR=rockot

Bug: 764576
Change-Id: I7ee8a26d350b65b196854fc9209a52fb607b580d
Reviewed-on: https://chromium-review.googlesource.com/923683
Reviewed-by: Cait Phillips <caitkp@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Tao Bai <michaelbai@chromium.org>
Reviewed-by: Ross McIlroy <rmcilroy@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#539558}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 31b8b4658644400222e359db415fb241c511ab76