S??n: Let NetworkContext configure Accept-Language, use that in place of ChromeHttpUserAgentSettings.
As for User-Agent, CreateDefaultNetworkContextParams() was already
filling in the correct thing, so nothing needed to be done with that,
though a redundant setting of it in IOThread::ConstructSystemRequestContext()
also got removed.

This required a patch from mmenke@ for StoragePartitionImpl::GetNetworkContext()
to avoid creating redundant NetworkContext's in some cases, so we don't
lose track of the ones that have HttpUserAgentSettings kept track of.

Cq-Include-Trybots: master.tryserver.chromium.linux:linux_mojo
Change-Id: Iee38daea39e60138abe90289bf9c4382c1da64ac
Reviewed-on: https://chromium-review.googlesource.com/956028
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#543562}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54248fa56e4a0845557dfd8a88340b515a9b4199