media: Use BindOnce everywhere in GMBVFP.
Changes the use of Bind and Closure to BindOnce and OnceClosure in
GpuMemoryBufferVideoFramePool.

It additionally introduces a trampoline in MailboxHoldersReleased so
that if the callback is called on media_thread it can skip going trough
the message loop.
Previous the callback would be bound to the media_thread loop.

Two DCHECKs have also be added to check thread expectations.

Change-Id: Ie704f6f6bdc4b7d9b8dd0f5ffec4675cd060d5e2
Reviewed-on: https://chromium-review.googlesource.com/959382
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#543609}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f35e389b9606f1b006005e83a3e8787ae8620f9b