Reland "media: Use BindOnce everywhere in GMBVFP."
Changes the use of Bind and Closure to BindOnce and OnceClosure in
GpuMemoryBufferVideoFramePool.

It additionally introduces a trampoline in MailboxHoldersReleased so
that if the callback is called on media_thread it can skip going trough
the message loop.
Previous the callback would be bound to the media_thread loop.

Two DCHECKs have also be added to check thread expectations.

TBR=dalecurtis, mcasas

Change-Id: I871e5505691b53a1871d2bd3e8a79ab1b75d2e69
Reviewed-on: https://chromium-review.googlesource.com/966583
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#543727}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c80fc114a7779e5e140caf7fe80d646eeca42fd9