[Picture in Picture] Include PiP in WMPI teardown.
This change notifies the WebMediaPlayerClient that
Picture-in-Picture has stopped when WMPI is deconstructed. The
client in question is <video>. With this change, the immediate
effect is the PiP interstitial will be hidden.

If the client is not in Picture-in-Picture mode, this is a no-op.

BUG: 726621
Change-Id: I3d240844ebed894a1f2566eaeae1a169be57a253
Reviewed-on: https://chromium-review.googlesource.com/954417
Commit-Queue: apacible <apacible@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#545380}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 302ad413f4902657ae83b4c142aa5a5824d5c2bf