media/gpu/vdatest: allow operation without GL
Make the --disable_rendering behave as it should, that is, disable
rendering altogether. With this patch, vdatest can run without any GL
support.

This has the advantage of reducing the subset of code tested, and to
remove false positives due to graphics regressions. It is also useful
during platform bringup where graphics are not always available and
working.

The thumbnails tests do require a renderer to create the thumbnails ;
therefore they are disabled when this option is set.

BUG=819542
TEST=Verified that vdatest was passing on kevin and hana, with both
rendering enabled and disabled and at various fps.

Change-Id: I119165e59cff24913918559d79a5c457e973cea4
Reviewed-on: https://chromium-review.googlesource.com/954770
Reviewed-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#546023}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3557c1a2021cced25249b1ca20a47238baeea5c0