Don't report video memory usage estimate when no frames are present.
preload=metadata suspend prevents decoding and rendering of the
first video frame. In these cases the video memory usage is zero,
but WMPI always includes an estimate for the frame held by the
compositor even if no frame is being held.

This artificially inflates the memory used in the preload=metadata
experiment. The fix is to not include this estimate if we've never
painted a frame before.

BUG=833564
TEST=new unittest.

Change-Id: I0b8811f04b27e5cc5c4ea5789cbd3d9247720771
Reviewed-on: https://chromium-review.googlesource.com/1014387
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#551167}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c2c5dcb1451ce1b2eaafc20419051776a0fdc881