Add FlingingRendererClientFactory
The FlingingRendererClientFactory (FRCF) is used when flinging URLs to
cast devices. The FRCF uses Mojo to create a media::Renderer in the
renderer process, which controls a FlingingRenderer in the browser
process.

The RendererFactorySelector selects the FRCF or not, conditionally on
whether FRCF has a valid presentation ID or not. This presentation ID
is retrieved from the blink::RemotePlayback object tied to WMPI,
through the RemotePlaybackClientWrapper.

N.B: A FlingingRendererClient class does not actually exist. We return
a MojoRenderer directly when CreateRenderer() is called, since a
FlingingRendererClient would just be a passthrough class.

Bug: 790766
Change-Id: Ief40c7b186ca35afb0e2fb9ee52bba993bcbf415
Reviewed-on: https://chromium-review.googlesource.com/1008905
Commit-Queue: Thomas Guilbert <tguilbert@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552553}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35a70ef8957c9e52441baec710acba03a7b0bd75