RELAND: use ScopedClosureRunner correctly in gpu_memory_buffer_video_frame_pool.
This CL relands the attached CL which got reverted due to a failing test
(WebRtcStressSourceSwitchBrowserTest.
MANUAL_SurvivesPeerConnectionSrcFeedSwitching on webrtc bots). The fundamental
cause for the failure was that the callback passed to ScopedClosureRunner was
getting called earlier than expected because the ScopedClosureRunner was
temporary.

The solution is to give a name to the ScopedClosureRunner object so that it
lives until the end of the function.

Changes can be seen here: crrev.com/c/1036162/1..2

Original CL description --------------------------------------------------------

This CL modifies the CopyRowsTo*Buffer functions to run the callback
using base::ScopedClosureRunner instead of manually running it at the
end.

Bug: 837243
Change-Id: I9ce269e32834f74bb594ddbc031ca9518e48ccad
Reviewed-on: https://chromium-review.googlesource.com/1036162
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#554972}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8fdba54d29d86114d6f5105893efabc9a4b22d44