media: Add MediaServiceTest.CdmProxy test
This tests CdmProxy creation and initialization using a ClearKeyCdmProxy
running in the MediaService. This is a prerequisite CL to add test for
Decryptor support in MediaService, which will be backed by the CdmProxy
running in the MediaService.

Bug: 806018
Test: More tests added
Change-Id: I9c5dd11eb5db5db3b553c7acfdba9c8120fd2489
Reviewed-on: https://chromium-review.googlesource.com/1038812
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#555563}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b971dfd641435863bfb872e1d2d97c09107eb1a