media: Remove NOTREACHED() in AndroidCdmFactory
Current we have a NOTREACHED() which essentially DCHECK that
MediaDrmBridge::IsKeySystemSupported(key_system) must return true. This
DCHECK is based on the fact that we called the same function when
deciding whether the key system is supported [1]. However, that is
called in the browser process and the DCHECK is in the GPU process.
Also, the DCHECK could happen much later. Further, the implemenation of
MediaDrmBridge::IsKeySystemSupported() actually calls into Java
MediaDrm, where it's hard to assume what it'll return.

This CL removes the NOTREACHED() so that if that happens we'll just
fail. UMA is also added to track how often this happens in the field.

[1] See CdmMessageFilterAndroid::OnQueryKeySystemSupport()

Bug: 843784
Change-Id: I6ce743033811e4d2bfbde75ccb715697e6205063
Reviewed-on: https://chromium-review.googlesource.com/1062803
Commit-Queue: Xiaohan Wang <xhwang@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#559358}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8d9220d54b7047d4223eb00856cdeeb36082c5fc