Fix for flaky track change test.
Omitting kNoClockless allowed the renderer to pull frames quickly rather
than trying to ratelimit them at the video fps. This was causing the
video to finish playback in the short timespan between being resumed and
suspended, causing the 'WaitUntilOnEnded' call to incorrectly return
early, as the 'ended_' attribute on the PipelineIntegrationTest had been
set to true. This ended_ boolean is only reset on a seek or a resume,
none of which had come after the suspend. This was flaking because
sometimes the video _wouldn't_ finish before the Suspend call, leaving
the pipeline test in an unended state.

Bug: 836785
Change-Id: I89d4382354e4cfcd6593b69f6cc02d6821492e41
Reviewed-on: https://chromium-review.googlesource.com/1072955
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#562008}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb593be96c0552c08d75c43076b569884874ba94