Add a optional lock for when AudioBus copies are desired.
WebAudioSourceProviderImpl is an unfortunate beast. It lives on
the render thread but may be accessed for 3+ threads (media, render,
audio). This really needs to be simplified or at least sectioned
better.

In the vast majority of cases, WASP is simply a passthrough
class that does nothing but stay out of the way. When copying or
WebAudio is attached things get complicated. In the copying case
an external caller will provide a callback that is used to copy
each AudioBus as its output.

The bug in this case is that writes to the callback were being
protected by a lock, but reads were only sometimes being protected
depending on if WebAudio was attached or not.

The fix is to introduce a new lock for the copy callback, but since
the vast majority of the time we don't need one, use an std::atomic
to avoid introduces yet-more audio fulfillment delay.

BUG=844833
TEST=none.

Change-Id: Idc29b65c383c99cad34072048435c5eff61eada3
Reviewed-on: https://chromium-review.googlesource.com/1073913
Reviewed-by: Olga Sharonova <olka@chromium.org>
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#562625}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 78923a9e70ee2f34290bd828ead3fdd91974a161