Let HTMLMediaElement drive playback state even for browser requests.
This is removing `PlaybackStateChanged()` from WebMediaPlayerClient and
instead adds `RequestPlay()` and `RequestPause()`. HTMLMediaElement will
notify the WebMediaPlayer when it states changes and there is no need
for the WebMediaPlayer to change its internal state and then notifies
the HTMLMediaEelement that things have changed. Indeed, it breaks some
assumptions that the HTMLMediaElement has.

The change also replaces some cast-related calls that are a by-product
of this design change.

Bug: 846490
Change-Id: Ibd60f571787e5f3a0d8a5ddfdbf4a35a2e53081e
Reviewed-on: https://chromium-review.googlesource.com/1074807
Reviewed-by: Dan Sanders <sandersd@chromium.org>
Reviewed-by: Philip Jägenstedt <foolip@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#562812}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 703106ed7990007156bc7d0656e0d139ff41966c