Reland 2 [Video Capture Service] Support accelerated jpeg decoding
This had been reverted because of a test failure on webrtc Win10 Tester for a
test that uses a real webcam.

Patchset 1 is the code as previously reviewed and reverted.
Patchset 2 adds a fix for the test failure.

Fix description:

The test failure revealed a mistake in the new code. Instead of operating the
Mojo connection to the jpeg decoder on the main task runner of the video capture
service, it would operate it from the thread the delivers the video frames from
the camera. On Windows, this thread is not a Chromium thread and as such does
not allow base::ThreadTaskRunnerHandle::Get(). The fix is to pass the correct
task runner to the factory code, see Patchset 2.

Original CL description:

* Move/rename class content::VideoCaptureGpuJpegDecoder to
  media::VideoCaptureJpegDecoderImpl and break dependencies to |content| in order
  to make it reusable for the video capture service. (I was unable to convince
  Gerrit that this is a move/rename. Sorry for this making it harder to review.)
* Split Gpu memory buffer factory methods from interface ui.mojom.Gpu to new interface
  ui.mojom.GpuMemoryBufferFactory
* Update class ui::ClientGpuMemoryBufferManager to depend on only the new interface
  ui.mojom.GpuMemoryBufferFactory.
* Inject dependencies on ui.mojom.GpuMemoryBufferFactory and viz.mojom.GpuService
  from Browser into video capture service when starting up the service.
* Enable the existing-but-disabled test cases for accelerated jpeg decoding in
  combination with video capture service

Design Doc: https://docs.google.com/document/d/169LS3U0DD7hHpqZSli0gpCaCqE1eWN4QXFSIGoruPwQ/edit?usp=sharing

Test: content_browsertests --gtest_filter=VideoCaptureBrowserTest.*
Bug: 820608, 720604
Change-Id: I6c77ce570aa8f0c3a54a12e0d3f14648c58005ac

TBR=avi@chromium.org,sadrul@chromium.org,jcliang@chromium.org,tsepez@chromium.org,sandersd@chromium.org,emircan@chromium.org,mojahsu@chromium.org,piman@chromium.org,chfremer@chromium.org

Change-Id: Id3493742537fa8a016dad391ee12e2869febf080
Reviewed-on: https://chromium-review.googlesource.com/1087867
Commit-Queue: Christian Fremerey <chfremer@chromium.org>
Reviewed-by: Christian Fremerey <chfremer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#564942}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 40cd31d1d8b7e496ca13227d8ae338e0f89b4061