Picture-in-Picture: handle frame deletion.
When an out of process iframe is deleted from the DOM, the player does
not notify the the browser process that it should leave the
Picture-in-Picture mode.

This is also fixing a TODO and updates a DCHECK. The TODO is because of
a refactor and the DCHECK was failing on a test that was just added.

Bug: 848044
Change-Id: I53ce77879cc08b70566449a705cdfb6723b4b013
Reviewed-on: https://chromium-review.googlesource.com/1095838
Reviewed-by: apacible <apacible@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#566411}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56f068d66db3c0943c8848eb16a21b966c8eca88