Resolve MediaCapabilities::decodingInfo promise even if mojo connection dies.
Promise for MediaCapabilities::decodingInfo() is resolved by
blink::WebMediaCapabilitiesQueryCallbacks which is passed to
WebMediaCapabilitiesClientImpl::DecodingInfo in //media/blink,
which, in some cases, means to resolve it in the callback
of an async mojo call. However, if the connections is lost meanwhile,
neither OnSuccess nor OnError is called and, in the result,
the promise resolver remains in pending state.

This patch makes sure the OnError is called when mojo connection
gets dropped.

Bug: 847211
Change-Id: I018792c56a3d3074b042e4318d9a8b85b3c16360
Reviewed-on: https://chromium-review.googlesource.com/1075308
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Commit-Queue: Chrome Cunningham <chcunningham@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#567874}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d6827fb7e8cabfcdc8b386f5a6f244f4c2e5743e